pyproject.toml
[project]
name = "newsletter-langgraph"
version = "0.1.0"
description = "Autonomous newsletter company with curation, billing, and marketing agents"
authors = [{ name = "Yudai Hatano", email = "yudaihatano.101@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain-ollama (>=0.3.6,<0.4.0)",
    "stripe-agent-toolkit (>=0.6.1,<0.7.0)",
    "stripe (>=12.3.0,<13.0.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langgraph (>=0.6.0,<0.7.0)",
    "playwright (>=1.54.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy[asyncio] (>=2.0.42,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "langgraph-sdk (>=0.2.0,<0.3.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.1"
notebook = "^7.4.4"
langgraph-cli = { extras = ["inmem"], version = "^0.3.6" }
langgraph-api = "^0.2.110"
pytest = "^8.4.1"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
